Add files via upload
This lab is a database application mainly focusing on the entities employee, department, and project. The main task was to allow the user to create a new project assignment with employee with the appropriate html files and python programs, while displaying the current project assignments on the home page. Finally, the user should be able to delete a given assignment.